split gui.h into .c/.h
